make quiz async
