+++ delay.js
@@ -1,0 +1,7 @@
+export function delay() {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve();
+    }, 200);
+  });
+}
