+++ index.html
@@ -8,6 +8,6 @@
       Start this exercise by adding a new .js file and include it in this file
       with type module:
     </p>
-    <script src="./index.js" type="module"></script>
+    <script src="./quiz.js" type="module"></script>
   </body>
 </html>
